work with NSKO; trying plot

+++ methods/func_load_from_file.py
@@ -8,7 +8,8 @@ def load_for_nsko(path):
         array_classes = [line.pop(-1) for line in lines]
 
         # print(lines, '\n', array_classes)
-        lines = [[int(number) for number in line] for line in lines]
+        # lines = [[int(number) for number in line] for line in lines]
+        lines = [[float(number) for number in line] for line in lines]
         array_classes = list(map(int, array_classes))
         return lines, array_classes
